refactor(viewer): extract useCameraTickSubscription hook
GizmoOverlay, WallEndpointOverlay, and SplitOverlay each ran an
identical RAF camera-tick loop (~20 lines apiece) to wake the
overlay on real viewport motion. The signature-string trick lets
the loop skip ticks when only frame number changed; missing that
optimisation in any one overlay would silently regress the gizmo
to per-frame re-render under a still camera.

Move the loop into apps/viewer/src/hooks/useCameraTickSubscription.ts:
the hook owns the RAF, the lastViewpoint signature ref, and the
frameTick state setter. Each overlay now passes
`(getViewpoint, active)` and discards the returned frameTick — the
internal setState alone re-renders the host, which re-runs the
post-memo projection. Trimmed the `frameTick` dependency from the
upstream `useMemo` blocks because camera motion doesn't actually
change those memoised values; the JSX projection refresh comes
from the host re-render itself.

Net 79 lines removed across three overlays for the new 63-line hook.
